almost finished InfoPessoa
quase acabando InfoPessoa, mudei coisa pouca nas outras classes: foi basicamente o tipo da variavel registro que as vezes estava int e as vezes long, deixei tudo long, e a variavel frequencia que estava as vezes double e as vezes float, deixei td double

+++ src/Aluno.java
@@ -5,7 +5,7 @@ public class Aluno extends Pessoa {
 	private int ocorr;
 	private Double[] notas;
 	
-	Aluno(long register, String nome, float freq, String turma, int ocorrencias, Double[] notas) {
+	Aluno(long register, String nome, double freq, String turma, int ocorrencias, Double[] notas) {
 		super(register, nome, freq);
 		this.turma = turma;
 		this.ocorr = ocorrencias;
